Add `CtxAttribute`s for source/binding specification to Parsing.py.
This makes the various meanings of visibility explicit in Ctx.  At the
moment, the meanings are all rolled back into a single string before
being passed to the Nodes.  The next step will be to pass the
CtxAttribute instances to Nodes directly, so Node implementations will
also be explicit about the meaning of visibility.

Passes runtests.py on my system.